feat(tilemap): :sparkles: Add event tooltip and refactor code
Very jank. Need to fix
