+++ src/graphics/primitives/sprite/sprite.wgsl
